(maint) Extract common test code into a BeakerUtils module
This lets us re-use the code in other tests.
